requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
supabase
openpyxl
plotly
fpdf2
matplotlib
